requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
Django==5.2.1
djangorestframework==3.16.0
idna==3.10
joblib==1.5.0
nltk==3.9.1
regex==2024.11.6
requests==2.32.3
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
